Move rlpx.rs to connection.rs
